# JS: Tokenize `:` as an operator