--- a/tests/languages/coffeescript/property_feature.test
+++ b/tests/languages/coffeescript/property_feature.test
@@ -5,9 +5,9 @@ foo_bar :
 
 [
 	["property", "foo"],
-	["punctuation", ":"],
+	["operator", ":"],
 	["property", "foo_bar"],
-	["punctuation", ":"]
+	["operator", ":"]
 ]
 
 ----------------------------------------------------