--- a/tests/languages/javascript/class-method_feature.test
+++ b/tests/languages/javascript/class-method_feature.test
@@ -58,10 +58,10 @@ class Test {
 	["parameter", [
 		["punctuation", "{"],
 		" props",
-		["punctuation", ":"],
+		["operator", ":"],
 		["punctuation", "{"],
 		" a",
-		["punctuation", ":"],
+		["operator", ":"],
 		" _A",
 		["punctuation", ","],
 		" b",