--- a/tests/languages/jsx/issue1235.test
+++ b/tests/languages/jsx/issue1235.test
@@ -14,7 +14,7 @@
 			["punctuation", "{"],
 			["punctuation", "{"],
 			" track",
-			["punctuation", ":"],
+			["operator", ":"],
 			["boolean", "true"],
 			["punctuation", "}"],
 			["punctuation", "}"]