--- a/tests/languages/jsx/issue1408.test
+++ b/tests/languages/jsx/issue1408.test
@@ -16,7 +16,7 @@
 			["punctuation", "{"],
 			["punctuation", "{"],
 			" marginLeft",
-			["punctuation", ":"],
+			["operator", ":"],
 			["template-string", [
 				["template-punctuation", "`"],
 				["interpolation", [