… tests/languages/javascript/operator_feature.test
… tests/languages/javascript/template-string_feature.test
… tests/languages/jsx/issue1335.test