This resolves #2072.

This will make JS and JSON consistent in how they tokenize `:`.