@mAAdhaTTah 2019-09-26
To make this broader: I'm not in love with changing all `:` from punctuation -> operator because of ternaries. Their usage in object literals is far more common than in ternaries, so punctuation is more likely to be correct.