@RunDevelopment 2019-09-26
I am for changing `:` to `operator` because of consistency. JSON also uses `operator` and it ternaries look a little strange. 
There is also the point that other highlighters also use something akin to `operator`, e.g. GitHub:

```js
const value = { hello: 'world', foo: (bar ? 1 : 0) };
```