@Golmote 2019-09-26
It's true that the JSON component already has it tagged as an operator for some reason.

I'm only worried that in the end, someone will open another issue asking for the exact opposite of this PR.