@MattMcFarland 2019-09-26
I agree that `:` should not be an operator when it is punctuation.  

I'm not sure how PrismJS is tokenizing things, but,  I believe that inorder to fix this, you would have to not broadly assume `:` is punctuation, but take in prior tokenized items as context and apply the the correct token type "punctuation" or "operator"