@RunDevelopment 2019-09-26
Well, IMO the underlying issue here is that there is no clear-cut definition of what differentiates an operator from mere punctuation.
I always felt that this divide is kind of artificial but it's quite useful for highlighting.

So my question: What defines an operator vs punctuation?
The JS language spec doesn't.