@MattMcFarland 2019-09-26
> It's true that the JSON component already has it tagged as an operator for some reason.
> 
> I'm only worried that in the end, someone will open another issue asking for the exact opposite of this PR.

I am afraid of this happening too.

The reason why I've brought this up in an issue is because our UX designer is looking at our code snippets on our site and not seeing parity github (we are using github's theme) - the first thing they pointed out to us was that our colons were not matching.  Since they are going over this with a fine-toothed comb, I'm sure if it does happen, it will happen much sooner rather than later lol.