@MattMcFarland 2019-09-26
> Well, IMO the underlying issue here is that there is no clear-cut definition of what differentiates an operator from mere punctuation.
> I always felt that this divide is kind of artificial but it's quite useful for highlighting.
> 
> So my question: What defines an operator vs punctuation?
> The JS language spec doesn't.

Having not looking at the spec myself, I believe you.  I think you have a very convincing argument for broadly assuming `:` is an operator.   Even if someone comes back saying they think it should be punctuation, it looks like `:` is used as an operator for github's syntax highlighting, if you look at this snippet:

```js
const foo = bar ? 0 : 1
const baz = {
  stuff: things,
  things: 'stuff',
  doStuff: () => MyThing.doStuff()
}