@MattMcFarland 2019-09-26
@Golmote Please.. Say it ain't so!  I'd much rather stick with prism if I can.  :) I do think that if there are no negative consequences for switching `:` to operator from punctuation, then you should do it.  It does look like it is treated like operator when looking at github snippets at least.