@RunDevelopment 2019-09-26
I also agree that we shouldn't change our tokenization to fit a specific theme.
I implemented this because of consistency.

---

@Golmote 
The spec even has [this punctuator section](https://www.ecma-international.org/ecma-262/6.0/#sec-punctuators).