@RunDevelopment 2019-09-26
@Golmote @mAAdhaTTah 
Do we have some guidelines to distinguish between punctuation and operator?

Because when I create a language, I just choose whatever I think looks better in Prism's themes. Even though this kinda goes against '_we shouldn't change our tokenization to fit a specific theme_'.