@MattMcFarland 2019-09-26
I agree you shouldnt change tokenization to fit a specific theme as well. 

When I bring up github's theme and am looking for parity, it's only to show my use case.  

As for the issue, I made it a point to leave theme's out of it and just stick with the facts, where an operator was being misread as punctuation.  That's what I was hoping a fix for :)

As for parity with Github, that's something my UX designer is using as a benchmark.  If they can give us some leeway that would be great, if not, it would just create more work on our board that we would probably have to defer.