@Golmote 2019-09-26
Consistency is a good enough argument to me, honestly. So I won't fight against this PR.

@RunDevelopment No we don't have guidelines for this. I personnally tried to stick to the specs, whenever possible, but I might have made aesthetic choices too, from time to time. Again, given Prism is no linter, that's perfectly fine.