@mAAdhaTTah 2019-09-26
I'm not sure that spec reference is that helpful, cuz it shows some tokens that are definitely operators, like spread `...`. Broadly, what is an operator is defined the spec. It sounds like punctuation is... made up? for the benefit of syntax highlighting?

@RunDevelopment So no, I don't think we have guidelines for "punctuation".

---

> It does look like it is treated like operator when looking at github snippets at least.

It looks this way because the GitHub theme highlights both with the same color. The `?` is also highlighted in that color, so hypothetically, if whatever highlighter GH uses highlighted them as different things but applied the same color, the result would be the same while "treating" them as different.