@mAAdhaTTah 2019-09-26
If you need to satisfy your designer, you could easily just change the color of punctuation & operator to be the same in your theme. No conflict 😄