@MattMcFarland 2019-09-26
@mAAdhaTTah I tried that but it didn't work.  It changes commas and dots to red (github treats them differently than operator)