@RunDevelopment 2019-09-26
@MattMcFarland Maybe we could add a bit functionality similar to [Highlight keywords](https://prismjs.com/plugins/highlight-keywords/) for general tokens to satisfy your UX designer? I'm thinking about the ability to modify the classes of tokens on a per-language level.
Anyway, we shouldn't discuss this here.

---

@Golmote 
> I personally tried to stick to the specs, whenever possible

Me too but most specs just call everything an operator that's not very applicable. Also, if present, what is punctuation vs operators varies from spec to spec.