@RunDevelopment 2019-09-26
> Ah so maybe GH's highlighter is treating : as an operator.

Seems like it. It's highlighted like `=` but differently from `(){}`.