@mAAdhaTTah 2019-09-26
Inconsistency between JS & JSON is a bit weird. We should probably bring them into alignment. The issue for me is it seems weird to apply "`:` is an operator" to JSON because it's definitely _not_ an operator there. I know this flies in the face of the original request, but my inclination is to make them both punctuation.