@Golmote 2019-09-26
Funnily enough, the Rouge syntax highlighter switched from having `?` and `:` tagged as operators to having them both tagged as punctuation in 2012.