@mAAdhaTTah 2019-09-26
lol there's your consistency: make `?` punctuation!