@RunDevelopment 2019-09-26
> make ? punctuation!

We can also do that. 
But with that, I have another question: What about [optional chaining](https://github.com/tc39/proposal-optional-chaining) and [nullish coalescing](https://github.com/tc39/proposal-nullish-coalescing)?

`??` is an operator for sure but `?.` ?