@RunDevelopment 2019-10-01
@Golmote 
I did some perf tests.

```md
sample        | func          | comp          | avg        | dev        | min        | max        | samples
------------- | ------------- | ------------- | ---------- | ---------- | ---------- | ---------- | ----------
prism-core.js | old operator  | 100% +-21%    | 0.07933    | 0.01674    | 0.04493    | 0.5169     | 1318
              | new operator  | 102% +-13%    | 0.08099    | 0.009933   | 0.04720    | 0.2069     | 1335
              | flat operator | 101% +-15%    | 0.08043    | 0.01168    | 0.04720    | 0.2379     | 1329
prism.js      | old operator  | 100% +-9%     | 3.470      | 0.3251     | 3.247      | 5.950      | 396
              | new operator  | 101% +-9%     | 3.503      | 0.3267     | 3.248      | 5.839      | 394
              | flat operator | 101% +-9%     | 3.495      | 0.3012     | 3.274      | 5.523      | 394
```

`avg` (average), `dev` (standard deviation), `min`, and `max` are all in ms. `comp` is the relative average and standard deviation compared to the lowest average.
`old operator` is the operator pattern before this PR, `new operator` is the operator pattern after this PR sans `:`, and `flat operator` is just all operators as a keyword list of sorts.
`prism-core.js` is what it sounds like from master, and `prism.js` is the current Prism version from the download page will all languages included.

<details>
<summary>Testing code</summary>

```js
const fs = require("fs");
const { performance } = require("perf_hooks");
// force the lazy init to happen
performance.now();


const samples = fs.readdirSync('./samples').filter(f => !/^_/.test(f)).map(f => {
	return {
		title: f,
		value: fs.readFileSync('./samples/' + f, 'utf8')
	}
});

benchmark(samples, [
	toTestCase(
		"old operator",
		/-[-=]?|\+[+=]?|!=?=?|<<?=?|>>?>?=?|=(?:==?|>)?|&[&=]?|\|[|=]?|\*\*?=?|\/=?|~|\^=?|%=?|\?|\.{3}/g
	),
	toTestCase(
		"new operator",
		/--|\+\+|\*\*=?|=>|&&|\|\||[!=]==|<<=?|>>>?=?|[-+*/%&|^!=<>]=?|\.{3}|[~?]/g
	),
	toTestCase(
		"flat operator",
		/--|-=|-|\+\+|\+=|\+|!==|!=|!|<<=|<<|<=|<|>>>=|>>>|>>=|>>|>=|>|===|==|=>|=|&&|&=|&|\|\||\|=|\||\*\*=|\*\*|\*=|\*|\/=|\/|~|\^=|\^|%=|%|\?|\.\.\./g
	),
]);

/**
 *
 * @param {string} title
 * @param {RegExp} regex
 */
function toTestCase(title, regex) {
	return {
		title,
		value: text => allMatches(regex, text),
	};
}


/**
 *
 * @param {{ title: string; value: T }[]} testSamples
 * @param {{ title: string; value: (value: T) => void}[]} testFunctions
 * @template T
 */
function benchmark(testSamples, testFunctions) {
	const warmupIterations = 1_000;
	const warmupMaxTime = 200;

	const warmupSample = testSamples[0].value;
	testFunctions.forEach(({ value }) => {
		const start = performance.now();
		for (let i = 0; i < warmupIterations; i++) {
			if (performance.now() - start > warmupMaxTime) {
				break;
			}
			value(warmupSample);
		}
	});

	const maxSamples = 10_000; // samples
	const maxTimePerFunc = 2_000; // ms


	const columnWidth = [
		Math.max(...testSamples.map(s => s.title.length)),
		Math.max(...testFunctions.map(s => s.title.length)),
	];

	function printLine(...cells) {
		let s = "";
		for (let i = 0; i < cells.length; i++) {
			let cell = String(cells[i]);
			if (i < columnWidth.length) {
				cell = cell.padEnd(columnWidth[i], " ");
			}
			if (s) {
				s += " | "
			}
			s += cell;
		}
		console.log(s);
	}

	let headerPrinted = false;

	/**
	 * @param {{ title: string; value: T }} sample
	 * @param {{ title: string; value: (value: T) => void}} func
	 * @param {Object<string, any>} result
	 */
	function printResult(sample, func, result) {
		if (!headerPrinted) {
			columnWidth.push(...Object.keys(result).map(k =>
				Math.max(10, k.length, Math.ceil(String(result[k]).length * 1.25))
			));
			printLine("sample", "func", ...Object.keys(result));
			printLine(...columnWidth.map(i => "-".repeat(i)));
			headerPrinted = true;
		}
		const cells = [
			func === testFunctions[0] ? sample.title : "",
			func.title
		];

		for (const key in result) {
			cells.push(result[key]);
		}
		printLine(...cells);
	}

	/**
	 * @param {Object<string, any>[]} results
	 */
	function printResults(sample, results) {
		const min = Math.min(...results.map(r => r.comp[0]));
		results.forEach(r => {
			const [avg, dev] = r.comp;
			r.comp = `${Math.round(100 * avg / min)}% +-${Math.round(100 * dev / min)}%`;
		});
		results.forEach((r, i) => {
			printResult(sample, testFunctions[i], r);
		})
	}

	for (const sample of testSamples) {
		const results = [];
		for (const func of testFunctions) {
			/** @type {number[]} */
			let samples = [];
			const samplingStart = performance.now();

			for (let i = 0; i < maxSamples; i++) {
				if (performance.now() - samplingStart > maxTimePerFunc) {
					break;
				}
				global.gc && global.gc();

				let start = performance.now();
				func.value(sample.value);
				samples.push(performance.now() - start);
			}

			const min = Math.min(...samples);
			const max = Math.max(...samples);
			const avg = samples.reduce((x, y) => x + y) / samples.length;
			const dev = Math.sqrt(samples.reduce((s, x) => s + (x - avg) ** 2) / (samples.length - 1))


			const prec = (n = 0) => n.toPrecision(4);

			results.push({
				comp: [avg, dev],
				avg: prec(avg),
				dev: prec(dev),
				min: prec(min),
				max: prec(max),
				samples: samples.length,
			});
		}
		printResults(sample, results);
	}
}

function allMatches(re, text) {
	if (!re.global) {
		throw new Error(`RegExp has to be global! ${re}`);
	}
	re.lastIndex = 0;
	while (re.exec(text)) { }
}
```

Run using Node.js v12.11.0 with the following command: `node --expose-gc bench.js`.
I take control over the GC (as much as I can) to avoid minor GCs during testing.

[bench.zip](https://github.com/PrismJS/prism/files/3678211/bench.zip)

</details>

Backtracking optimization or not doesn't seem to matter. 
While your pattern generally seems to be just a tat faster than the other ones (like 1% faster), it's hard to tell because of the noise.

To reduce this noise, I experimented a little with v8's GC and got it down by ~10% but there's still too much to conclusively say that your pattern is faster.